Rename system strings to constants
