+++ examples/medplum-photon-integration/src/bots/constants.ts
