+++ examples/medplum-photon-integration/src/components/PatientPrescription.tsx
@@ -5,7 +5,7 @@ import { Patient } from '@medplum/fhirtypes';
 import { Document, useMedplum } from '@medplum/react';
 import { IconCircleCheck, IconCircleOff } from '@tabler/icons-react';
 import { useState } from 'react';
-import { NEUTRON_HEALTH_BOTS, NEUTRON_HEALTH_PATIENTS } from './../bots/system-strings';
+import { NEUTRON_HEALTH_BOTS, NEUTRON_HEALTH_PATIENTS } from '../bots/constants';
 
 interface PatientPrescriptionProps {
   patient: Patient;
